src/stralloc.rs: Consolidate from Corroded `stralloc_*.rs`
Consolidates all Corrode-translated source code for stralloc-related
functionality into a common `stralloc` module.

Moves all `stralloc_*()` into an `impl stralloc` block, as these
functions should eventually all accept `self`.

Code is otherwise unchanged from the original Corrode output.
