--- src/stralloc_cats.rs
@@ -1,24 +1,0 @@
-use libc;
-
-extern "C" {
-    fn stralloc_catb(arg1: *mut stralloc, arg2: *const u8, arg3: u32) -> i32;
-}
-
-#[derive(Copy)]
-#[repr(C)]
-pub struct stralloc {
-    pub s: *mut u8,
-    pub len: u32,
-    pub a: u32,
-}
-
-impl Clone for stralloc {
-    fn clone(&self) -> Self {
-        *self
-    }
-}
-
-#[no_mangle]
-pub unsafe extern "C" fn stralloc_cats(mut sa: *mut stralloc, mut s: *const u8) -> i32 {
-    stralloc_catb(sa, s, libc::strlen(s as *const i8) as u32)
-}
